Agregada info a los .txt / clases ordenadas en paquetes
