+++ src/main/java/com/jeaguirre/pronosticosdeportivos/enumeraciones/EnumEquipo.java
@@ -3,12 +3,12 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package com.jeaguirre.pronosticosdeportivos;
+package com.jeaguirre.pronosticosdeportivos.enumeraciones;
 
 /**
  *
  * @author aguir
  */
-public enum EnumLocalia {
-    LOCAL, VISITANTE;
+public enum EnumEquipo {
+    EQUIPO1, EQUIPO2, EQUIPO3, EQUIPO4;
 }
