+++ src/main/java/com/jeaguirre/pronosticosdeportivos/jugador/JugadorApuesta.java
@@ -3,7 +3,10 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package com.jeaguirre.pronosticosdeportivos;
+package com.jeaguirre.pronosticosdeportivos.jugador;
+
+import com.jeaguirre.pronosticosdeportivos.enumeraciones.EnumLocalia;
+import com.jeaguirre.pronosticosdeportivos.enumeraciones.EnumResultado;
 
 /**
  *
