+++ src/main/java/com/jeaguirre/pronosticosdeportivos/torneo/Torneo.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package com.jeaguirre.pronosticosdeportivos;
+package com.jeaguirre.pronosticosdeportivos.torneo;
 
 import java.util.HashMap;
