+++ src/main/resources/pronostico.txt
@@ -1,2 +1,32 @@
 Julián,ronda1,partido1,LOCAL,GANADOR
 Julián,ronda1,partido2,LOCAL,EMPATE
+Julián,ronda2,partido1,LOCAL,GANADOR
+Julián,ronda2,partido2,LOCAL,EMPATE
+Julián,ronda3,partido1,LOCAL,GANADOR
+Julián,ronda3,partido2,LOCAL,EMPATE
+Julián,ronda4,partido1,LOCAL,GANADOR
+Julián,ronda4,partido2,LOCAL,EMPATE
+Sebastián,ronda1,partido1,LOCAL,EMPATE
+Sebastián,ronda1,partido2,LOCAL,EMPATE
+Sebastián,ronda2,partido1,LOCAL,EMPATE
+Sebastián,ronda2,partido2,LOCAL,EMPATE
+Sebastián,ronda3,partido1,LOCAL,EMPATE
+Sebastián,ronda3,partido2,LOCAL,EMPATE
+Sebastián,ronda4,partido1,LOCAL,EMPATE
+Sebastián,ronda4,partido2,LOCAL,EMPATE
+María,ronda1,partido1,LOCAL,GANADOR
+María,ronda1,partido2,LOCAL,GANADOR
+María,ronda2,partido1,LOCAL,GANADOR
+María,ronda2,partido2,LOCAL,GANADOR
+María,ronda3,partido1,LOCAL,GANADOR
+María,ronda3,partido2,LOCAL,GANADOR
+María,ronda4,partido1,LOCAL,GANADOR
+María,ronda4,partido2,LOCAL,GANADOR
+Frida,ronda1,partido1,LOCAL,GANADOR
+Frida,ronda1,partido2,LOCAL,GANADOR
+Frida,ronda2,partido1,LOCAL,GANADOR
+Frida,ronda2,partido2,LOCAL,GANADOR
+Frida,ronda3,partido1,LOCAL,GANADOR
+Frida,ronda3,partido2,LOCAL,GANADOR
+Frida,ronda4,partido1,LOCAL,GANADOR
+Frida,ronda4,partido2,LOCAL,GANADOR
